Move config objects to config package
