+++ robotpy_build/config/autowrap_yml.py
@@ -454,7 +454,7 @@ class TemplateData(Model):
     doc_append: Optional[str] = None
 
 
-class HooksDataYaml(Model):
+class AutowrapConfigYaml(Model):
     """
     Format of the file in [tool.robotpy-build.wrappers."PACKAGENAME"]
     generation_data
