+++ robotpy_build/config/pyproject_toml.py
@@ -2,7 +2,6 @@
 # pyproject.toml
 #
 
-import os
 import re
 from typing import Dict, List, Optional
